requirements.txt
elasticsearch==8.10.0
nltk==3.8.1
pandas==2.0.3
numpy==2.2.2
scipy==1.15.1
pytrec_eval==0.5
sentence-transformers==2.6.1
faiss-cpu==1.7.1